feat: create BillableBuilderService
