+++ billable/Cargo.toml
@@ -8,3 +8,6 @@ edition = "2021"
 [dependencies]
 common = { path = "../common" }
 chrono = "0.4.26"
+async-trait = "0.1.51"
+log = "0.4.18"
+anyhow = "1.0.71"
